feat: general ui overhaul, exported switchcolortheme added switchcolormode to editor page added margins added logo to footer in menu switchcolortheme button now in @components/header/SwitchColorTheme.tsx removed unnecessary imports
