+++ lumium-api/.env.development
@@ -1,2 +1,5 @@
 API_HOST="http://localhost:5000"
 SPACE_HOST="http://localhost:3000"
+SUPERTOKENS_API_KEY=ABLgqizyFjHwKDfGfkxAQhbDAQXmRw
+SUPERTOKENS_CONNECTION_URI=https://57082451e0d511ecae81af738acb1c05-eu-west-1.aws.supertokens.io:3567/
+DATABASE_URL="postgres://development:development@localhost:5432/lumium"
