+++ lumium-space/pages/page/[[...path]].tsx
@@ -1,12 +1,14 @@
 import { Authenticator } from "@security";
 import { LumiumRenderer } from "@components";
 import { TopDrawerMenu } from "@components/menu";
-
+import { Box } from '@chakra-ui/react'
 const Space: React.FC = () => {
     return (
         <Authenticator>
             <TopDrawerMenu />
-            <LumiumRenderer />
+            <Box m={["1%"]}>
+                <LumiumRenderer />
+            </Box>
         </Authenticator>
     );
 };
