+++ lumium-space/src/components/headers/index.tsx
@@ -1,2 +1,3 @@
 export { Header } from "./Header";
 export { ProfileMenu } from "./ProfileMenu";
+export { SwitchColorTheme } from "./SwitchColorTheme";
